fix: properly propagate AbortError and reset camera on animation stop
When the Stop button was clicked during animation playback, the AbortError
was being caught in useAnimationController but then silently swallowed instead
of being re-thrown. This prevented the play handler in visual.tsx from detecting
the abort and resetting the camera to the start position.

Changes:
- Remove unnecessary try/catch in useAnimationController.play that was
  swallowing AbortError - let it propagate to calling code
- Add camera reset in visual.tsx abort handler to reset to start position
  when animation is stopped
- Update createRAFAnimation to properly accept and handle AbortSignal for
  RAF-based animations (playFollowPath, playRotateAroundPoint)
- Improve error logging to not log expected AbortErrors

This ensures that when Stop is clicked:
1. Animation aborts cleanly
2. Camera resets to start position
3. No console errors are logged
